Complete Test Reorganization and GitHub Actions Workflow Unification (#22)
* fix: Use coreclr runtime by default for ADOMD.NET compatibility

- Changed default pythonnet runtime from netfx to coreclr on Windows
- This fixes integration tests that were failing due to ADOMD.NET not being available
- All tests now pass (32 passed, 2 skipped)
- coreclr runtime works properly with NuGet packages used for ADOMD.NET

* perf: Significantly improve SSE endpoint test performance

- Fixed server startup detection using socket connection test instead of HTTP requests
- Server startup time reduced from ~75s to ~4s
- Added explicit --port argument to subprocess call
- Improved error handling and debugging output
- Test now uses proper server persistence (MCP_PERSIST=1)
- Socket-based connectivity check avoids SSE endpoint timeout issues

* docs: Add comprehensive testing requirements to AGENTS.md

Key additions:
-  MANDATORY: Full test suite validation before task completion
- Clear prohibition against deleting failing tests without approval
- Layered test execution approach (unit  local  integration)
- Explicit test markers and organization guidelines
- Updated validation commands for new test structure
- Emphasis that ALL tests must pass before committing

This ensures no functionality is broken and maintains code quality standards.

* Reorganize tests into layered architecture

- Create unit/, local/, integration/ test directories
- Move tests to appropriate layers based on dependencies:
  * unit/: No external dependencies (18 tests)
  * local/: Require local server startup (2 tests)
  * integration/: Require external services (14 tests, 2 skipped)
- Fix import paths in local tests using os.path.dirname(__file__)
- Add __init__.py files for each test layer
- Update pytest.ini with integration marker
- Fix ADOMD.NET loading by removing SKIP_ADOMD_LOAD environment variable
- All test layers now pass: 32 passed, 2 skipped in 58.80s

Tests can now be run by layer:
- pytest tests/unit/ (fast unit tests)
- pytest tests/local/ (server startup tests)
- pytest tests/integration/ (full integration tests)
- pytest tests/ (all layers)

* perf: Optimize test_server_stays_running performance

- Remove HTTP request-based server detection (unreliable for MCP)
- Use simple process liveness check instead of socket connection
- Reduce test duration from ~24s to ~5s (79% improvement)
- Simplify logic: just verify server process doesn't die within 5 seconds
- Remove unused httpx import
- Add explicit --port argument for clarity

Local test layer now runs in ~9s total (was ~28s)
Full test suite: 32 passed, 2 skipped in 44.79s

* feat: Complete test reorganization and GitHub Actions workflow unification

 Test Architecture Reorganization:
- Reorganized tests into layered structure (unit/local/integration)
- Unit tests: Run on multiple Python versions (3.10, 3.11, 3.12) - fast, no external deps
- Local tests: Run on Python 3.11 - require server startup
- Integration tests: Run on Python 3.11 - require external services
- Optimized test_server_stays_running from ~24s to ~5s using process monitoring
- All 32 tests pass with improved performance

 GitHub Actions Workflow Unification:
- Created shared composite action: .github/actions/setup-powerbi-environment/
- Unified environment setup across Docker, Copilot, and CI using install_dotnet_adomd.sh
- Standardized Python 3.11 across all environments for consistency
- Simplified copilot-setup-steps.yml to use composite action
- Updated tests.yml with layered test execution strategy
- Removed merge commit exclusion - tests now run after PR merges too

 Key Benefits:
- Environment consistency between Copilot and CI/CD
- Optimized test execution with dependency flow
- Eliminated code duplication via composite actions
- Better performance with targeted test layers

* Update tests/local/test_server_process.py


* Fix code formatting issues and relocate agent instructions

Code Formatting Fixes:
- Fix argument formatting in test_server_process.py (--port parameter)
- Fix whitespace issues in test_sse_endpoint.py (trailing spaces, socket import)
- Fix string quotes in test_connector.py (double quotes for path)
- Fix f-string without placeholders in test_sse_endpoint.py (flake8 compliance)

Agent Instructions Relocation:
- Move AGENTS.md to .github/copilot-instructions.md
- Update documentation with critical formatting requirements
- Emphasize mandatory black formatting checks before commits
- Add quick validation commands for CI/CD compliance

All formatting now passes black --check validation required by GitHub Actions.
All 32 tests pass (18 unit + 2 local + 12 integration).

* Enable integration tests on pull requests

- Remove conditional execution of integration tests (was push to master only)
- Integration tests now run on both pull requests and push to master
- Maintains layered test execution: unit -> local -> integration

* Fix ENABLE_INTEGRATION_TESTS to use variables instead of secrets

- Change ENABLE_INTEGRATION_TESTS from secrets to vars context
- Variables are more appropriate for non-sensitive configuration flags
- Secrets remain for sensitive data like API keys and credentials

* Add copilot environment to integration tests

- Use environment: copilot in integration-tests job
- Ensures access to ENABLE_INTEGRATION_TESTS variable and all secrets
- Matches configuration used in copilot-setup-steps.yml
- Should resolve integration tests not running in Copilot environment

* Fix TEST_MIN_TABLES_COUNT to use variables instead of secrets

- Change TEST_MIN_TABLES_COUNT from secrets to vars context
- Variables are appropriate for non-sensitive configuration values
- Should resolve ValueError: invalid literal for int() with base 10: '' error
- Configuration values vs sensitive credentials properly separated

---------
