+++ pytest.ini
@@ -11,6 +11,7 @@ addopts =
     --color=yes
 markers =
     integration: marks tests as integration tests (may require external services)
+    local: marks tests as local tests requiring server startup  
     unit: marks tests as unit tests
     slow: marks tests as slow running
 filterwarnings =
