+++ tests/integration/__init__.py
@@ -1,0 +1,1 @@
+# Integration tests - require external services (Power BI, OpenAI)
