+++ tests/local/__init__.py
@@ -1,0 +1,1 @@
+# Local tests - require server startup but no external services
