+++ tests/unit/test_openai_optional.py
@@ -7,7 +7,7 @@ import sys
 import pytest
 from mcp import types as mcp_types
 
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "src"))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "..", "src"))
 
 from server import PowerBIMCPServer
